# perf: pre-compute metric keys