--- a/emasamplerate_test.go
+++ b/emasamplerate_test.go
@@ -533,3 +533,128 @@ func TestEMASampleRate_Start(t *testing.T) {
 		})
 	}
 }
+
+func TestEMASampleRate_GetMetrics(t *testing.T) {
+	tests := []struct {
+		name           string
+		prefix         string
+		requestCount   int64
+		eventCount     int64
+		currentCounts  map[string]float64
+		burstCount     int64
+		intervalCount  uint
+		existingPrefix string
+		expectedResult map[string]int64
+		expectNil      bool
+	}{
+		{
+			name:          "first call with prefix",
+			prefix:        "test_",
+			requestCount:  100,
+			eventCount:    500,
+			currentCounts: map[string]float64{"key1": 10, "key2": 20},
+			burstCount:    5,
+			intervalCount: 3,
+			expectedResult: map[string]int64{
+				"test_request_count":  100,
+				"test_event_count":    500,
+				"test_keyspace_size":  2,
+				"test_burst_count":    5,
+				"test_interval_count": 3,
+			},
+		},
+		{
+			name:          "empty prefix",
+			prefix:        "",
+			requestCount:  42,
+			eventCount:    84,
+			burstCount:    0,
+			intervalCount: 1,
+			currentCounts: map[string]float64{"key1": 5},
+			expectedResult: map[string]int64{
+				"request_count":  42,
+				"event_count":    84,
+				"burst_count":    0,
+				"interval_count": 1,
+				"keyspace_size":  1,
+			},
+		},
+		{
+			name:          "zero counts",
+			prefix:        "zero_",
+			currentCounts: map[string]float64{},
+			expectedResult: map[string]int64{
+				"zero_request_count":  0,
+				"zero_event_count":    0,
+				"zero_keyspace_size":  0,
+				"zero_burst_count":    0,
+				"zero_interval_count": 0,
+			},
+		},
+		{
+			name:           "same prefix second call",
+			prefix:         "same_",
+			requestCount:   200,
+			eventCount:     1000,
+			currentCounts:  map[string]float64{"key1": 15},
+			burstCount:     10,
+			intervalCount:  2,
+			existingPrefix: "same_",
+			expectedResult: map[string]int64{
+				"same_request_count":  200,
+				"same_event_count":    1000,
+				"same_keyspace_size":  1,
+				"same_burst_count":    10,
+				"same_interval_count": 2,
+			},
+		},
+		{
+			name:           "different prefix returns nil",
+			prefix:         "new_",
+			requestCount:   150,
+			eventCount:     750,
+			currentCounts:  map[string]float64{"key1": 25},
+			existingPrefix: "old_",
+			expectNil:      true,
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			a := &EMASampleRate{
+				requestCount:  tt.requestCount,
+				eventCount:    tt.eventCount,
+				currentCounts: tt.currentCounts,
+				burstCount:    tt.burstCount,
+				intervalCount: tt.intervalCount,
+			}
+
+			// Set existing prefix if specified
+			if tt.existingPrefix != "" {
+				a.prefix = tt.existingPrefix
+				a.requestCountKey = a.prefix + requestCountSuffix
+				a.eventCountKey = a.prefix + eventCountSuffix
+				a.keyspaceSizeKey = a.prefix + keyspaceSizeSuffix
+				a.burstCountKey = a.prefix + burstCountSuffix
+				a.intervalCountKey = a.prefix + intervalCountSuffix
+			}
+
+			result := a.GetMetrics(tt.prefix)
+
+			if tt.expectNil {
+				assert.Nil(t, result)
+				return
+			}
+
+			assert.NotNil(t, result)
+			assert.Equal(t, tt.expectedResult, result)
+
+			assert.Equal(t, tt.prefix, a.prefix)
+			assert.Equal(t, tt.prefix+requestCountSuffix, a.requestCountKey)
+			assert.Equal(t, tt.prefix+eventCountSuffix, a.eventCountKey)
+			assert.Equal(t, tt.prefix+keyspaceSizeSuffix, a.keyspaceSizeKey)
+			assert.Equal(t, tt.prefix+burstCountSuffix, a.burstCountKey)
+			assert.Equal(t, tt.prefix+intervalCountSuffix, a.intervalCountKey)
+		})
+	}
+}