--- a/totalthroughput.go
+++ b/totalthroughput.go
@@ -48,8 +48,12 @@ type TotalThroughput struct {
 	lock sync.Mutex
 
 	// metrics
-	requestCount int64
-	eventCount   int64
+	requestCount    int64
+	eventCount      int64
+	prefix          string
+	requestCountKey string
+	eventCountKey   string
+	keyspaceSizeKey string
 }
 
 // Ensure we implement the sampler interface
@@ -174,10 +178,21 @@ func (t *TotalThroughput) LoadState(state []byte) error {
 func (t *TotalThroughput) GetMetrics(prefix string) map[string]int64 {
 	t.lock.Lock()
 	defer t.lock.Unlock()
-	mets := map[string]int64{
-		prefix + "request_count": t.requestCount,
-		prefix + "event_count":   t.eventCount,
-		prefix + "keyspace_size": int64(len(t.currentCounts)),
+
+	if t.prefix == "" {
+		t.prefix = prefix
+		t.requestCountKey = t.prefix + requestCountSuffix
+		t.eventCountKey = t.prefix + eventCountSuffix
+		t.keyspaceSizeKey = t.prefix + keyspaceSizeSuffix
 	}
-	return mets
+
+	if t.prefix != prefix {
+		return nil // if the prefix doesn't match, return nil
+	}
+
+	return map[string]int64{
+		t.requestCountKey: t.requestCount,
+		t.eventCountKey:   t.eventCount,
+		t.keyspaceSizeKey: int64(len(t.currentCounts)),
+	}
 }