## Which problem is this PR solving?

Generating the same metric name on each call to GetMetrics creates unnecessary heap allocation. This PR memoize the result after the first call to allow reuse of those keys

## Short description of the changes

- add metric key names to the Sampler struct
- memoize the result after first call
- add tests and benchmarks

I didn't change the GetMetrics signature to avoid breaking changes to the library

## Benchmark result
```
**OLD**
goos: darwin
goarch: arm64
pkg: github.com/honeycombio/dynsampler-go
cpu: Apple M2 Max
BenchmarkEMAThroughputGetMetrics-12    	51280945	      218.2 ns/op	    376 B/op	      7 allocs/op
PASS
ok  	github.com/honeycombio/dynsampler-go	11.640s

**NEW**
goos: darwin
goarch: arm64
pkg: github.com/honeycombio/dynsampler-go
cpu: Apple M2 Max
BenchmarkEMAThroughputGetMetrics-12    	94783080	      109.9 ns/op	    256 B/op	      2 allocs/op
PASS
ok  	github.com/honeycombio/dynsampler-go	11.092s

```