@VinozzZ 2025-07-07
> I think this is as good as we can do without changing the interface. But now I feel bad about the interface and think that maybe someday we'll want to go to constants rather than supporting prefixes.

One day we will commit to a breaking change and go to constants. I think this is good enough for now